requirements.txt
python-telegram-bot
yt-dlp
instagrapi
pillow